requirements.txt
websockets
aiohttp
pandas
numpy
scipy
scikit-learn
streamlit
matplotlib
python-dotenv
pyyaml
loguru
pytest
